Trigger events during the paste process
These act as hooks to allow for means to debug this conversion process.

There are 3 events:

htmlinput: provides details of the HTML (if any) that was received,
always triggered.
sanitise: provides details of the HMTL once sanitised, only fires if
HTML is input.
govspeak: provides details of the converted govspeak, only fires if HTML
is input.

As IE11 does not support the CustomEvent constructor we use a snippet from the
polyfill provided on MDN [1]. If the usage of these events is to
increase in complexity we should use a polyfill from NPM.

[1]: https://developer.mozilla.org/en-US/docs/Web/API/CustomEvent/CustomEvent#Polyfill