Re-instate IE11 fix
D'oh! I somehow lost this in a bad merge.

To try reduce the chances of this happening again I've added a test for
this. By some witchcraft I can mock an import after running other
imports - hopefully that's intended functionality. This allows mocking
the htmlFromHiddenElement function which simplifies this test a lot.