Make HTML easier to read in debug
This uses the [js-beautify library][1] to style the HTML that is
debugged to make it easier to read. It still has style attributes which
are the hardest part to read. I'm not too sure of a simple way to remove
them though or if we even should.

It sets a pre-wrap on the pre elements so they're easier to read without
scroll. However as HTML has massive style attributes and wrapped
prettifying this is still white-space: pre for scrolling.

[1]: https://github.com/beautify-web/js-beautify