Rename the htmlFromHiddenElement module
This names this as legacyHtmlFromPaste to be easier for readers of the
codebase to understand.

I've also removed the check as to whether window.clipboardData exists
before using the legacyHtmlFromPaste method since it should be fine to
use this even if window.clipboardData doesn't exist (as we don't use it)
and if we do actually need it for non IE browsers then we should do the
check inside the legacyHtmlFromPaste module.