Auto-tag and publish to NPM via Travis
This configures Travis to automatically push a new tagged version when
the package version in package.json changes, as well as push the new
version to NPM if it isn't already there. We use different checks for
each function to ensure they can run independently if one of them fails.