Clean up nested links
This attempts to clean up cases where a user has copied and pasted markdown
which contains an HTML link. Normally Turndown would convert the `a` element to
markdown, resulting in a nested markdown link e.g. `[link
markdown]([https://www.gov.uk/](https://www.gov.uk/))`. As a post processing
step we scan for instances of nested markdown, capture the second URL and
replace the nested markdown with the URL.