Clean up nested links
This attempts to clean up cases where a user has copied and pasted markdown
which contains an HTML link (this is usually as a result of Google docs
autolinking URLs). Normally Turndown would convert the `a` element to markdown,
resulting in a nested markdown link e.g. `[link
markdown]([https://www.gov.uk/](https://www.gov.uk/))`. This filters for `a`
elements and checks that the preceding text ends with `](`, which denotes that
the `a` tag is sitting within the URL part of the link markdown. Then we
replace the `a` element with its href value.

We are expecting that when this situation occurs the text part of the link
markdown will sit within a `span` element, as that is what we have observed so
far in testing. However, we want to ensure that this works even if the text
part of the link markdown is not in an element, hence the additional test case.