Fix incorrect HTML for ordered lists produced by Google Docs
As with <ul> elements Google Docs also produced incorrect ordered lists
where the <ol> elements are siblings of <li> elements rather than
children. The fix for this is much the same as for unordered lists.

There is however a problem that the Google markup then breaks the
ordering of lists so they are off by one because turndown only expects a
ol to have children of type li (reasonably). To resolve this I have
ported the list item method from turndown so that we can fix it with a
change to the method.