Update the list indenting to match GOV.UK style
Since we now handle the list markdown rather than turndown we have the
opportunity to update the indenting to be more consistent with GOV.UK
style. This sets the lists to have just one space after a marker and 3
spaces before it.

I wrestled for a bit trying to work out what should be the correct level
of indentation. Some places advise 2 characters, others 3 and others 4.
If you are working in monotype fonts it's most readable to have 2 for
ul's and 3 for ol's. As we expect users not to be typing in monotype
fonts this isn't too big a concern and 3 seems to offer a nice balance
of readability and without special cases for each list type.