Expand the public API
This adds another function to the public API of this module,
htmlToGovspeak, which is a single function that will convert given HTML
into Govspeak. Previously this was a private method.

The reason for adding this to the API is that has been a difficult
process to test out quite how HTML is converted to Govspeak solely by a
private API - which normally meant just unit tests. The intention of
doing this is to support another example file that will allow inputting
HTML as text and seeing the Govspeak outputted.

Now in an ideal world this would be just another function exposed with
the paste listener being the default. However this doesn't fly in the
current UMD world we're in, so I've taken the approach of exposing both
functions on the object and recommending users add the listener in their
code - which is arguably clearer and more extensible.

I decided the exposed function made more sense as htmlToGovspeak than
just toGovspeak so I've re-named this across the package.

So the conclusion of this is that importing paste-html-to-govspeak now
allows two functions that can be imported pasteListener and
htmlToGovspeak. I pondered whether listener was better than
pasteListener since that is implied by the package name, but it just
didn't feel very clear as listener.

I did wonder if we end up with redundant terminologies going on here
where this whole package could be renamed htmlToGovspeak and then have
methods of pasteListener and convert exported. That seemed a pretty
major change and we still anticipate the main use case to be for paste
events.