Combine example files with new input HTML functioanlity
In testing this package there was a pain point in the difficultly of
seeing how particular HTML would convert to markdown. I decided to
remedy this with another example, although adding that pushed us up to 3
examples so I then felt it was better to consolidate to a single one.

So this commit consolidates the debug.html and index.html files into a
single file and also adds a HTML input to govspeak functionality.

The HTML to Govspeak functionality listens for input events on a
textarea and then outputs the govspeak in a pre element below. It also
runs when the page is initially loaded in case there is already a value
of the field (e.g. if you press back on your browser).