Convert headers to h2 or h3 rather than strip
GOV.UK supports h2 and h3 headers (with h4 allowed but discouraged). h1
headers are reserved for the page title. Previously this was supported
by stripping h1, h4, h5, and h6 headers.

A product decision has been made that instead of stripping h1, h4 and h5
elements these are converted to h2 or h3s, so that they are preserved of
headers but lose some of their hierarchical meaning.