Convert Microsoft Word lists to markdown lists
This parses the HTML that Microsoft Word produces for lists and
converts this to markdown.

It has to rely on parsing class names and style attributes, so is very
fragile for any iterations that Microsoft make to this. Hopefully future
versions may use actual HTML list markup so this may not be needed.

Lists are presented in a series of paragraphs with the first item having
a class of "MsoListParagraphCxSpFirst", the last item having a class of
"MsoListParagraphCxSpLast" and all the other list items having a class
of "MsoListParagraphCxSpMiddle". This behaviour is regardless of whether
there are nested lists or not.

Whether an item is nested in a list or not can be determined by a style
property of "mso-list". This contains a value of level and an integer to
determine the nesting of the item. It also has different values either
side of the level that I've not been able to decipher, but I think they
relate to the number of lists in a document.

Inside a list is a span element with a style of "mso-list:Ignore" this
element contains the list marker. The only way I've been able to
determine if we're in an ordered or unordered list is to check for the
presence of a period in this value.