Widen support of Word HTML lists
So shortly after releasing the Word list HTML conversion I found some
documents opened in word that didn't behave the same as the other ones I
tested (le sigh).

Looking at the HTML that was produced there showed a lack of list class
names and instead usage of the MsoNormal class (which Word normally uses
for paragraphs).

It wasn't too hard to change the code we have now to handle both of
these scenarios so I've updated it for this.