Manually strip non-breaking spaces from MS word comment nodes
This deals with a part of turndown which is somewhat inelegant, if we
remove an element that starts or ends with a non-breaking space it will
be interpreted as trailing or leading whitespace of the node and
included in output. Rather than removed as one might expect.

For example:

```
const service = new TurndownService()
service.remove('span')
service.turndown('test <span><span><span>a &nbsp;</span>b &nbsp;</span>c &nbsp;</span>test')
=> 'test   test'
service.turndown('test <span><span><span>a</span>b</span>c</span>test')
=> 'test test'
service.turndown('test <span><span><span>a </span>b </span>c </span>test')
=> 'test  test'
```

This seems to be intentional changes to Turndown as per:
https://github.com/mixmark-io/turndown/commit/27bdac0c2fecda9135a92bc5f4fe20ed25831309
though it is quite confusing.

This seems to be the case with Microsoft word comments that we want to
remove entirely, however they end up with a trailing non-breaking space
character.

This code adjusts the node at the point of conversion to remove any nbsp
whitespace. It wouldn't surprise me hugely if there were other problems
related to this that occur, but I can't think of any.