Ensure table plugin works for tables without a correctly formulated header row
Whilst technically HTML tables should contain a THEAD element with TH for each cell inside the header, most programs tested (including Google Docs, LibreOffice and Google Sheets) do not produce such tables by default. Instead the first row will still be a standard TR element with TD elements for each cell. This commit modifies the table handling code so that it will still work for incorrectly formatted tables. It assumes that the first row of the table can be used as the heading row and formats the markdown appropriately.